tests/wpt/web-platform-tests/shadow-dom/elements-and-dom-objects/the-content-html-element/test-001.html

<!DOCTYPE html>
<!--
Distributed under both the W3C Test Suite License [1] and the W3C
3-clause BSD License [2]. To contribute to a W3C Test Suite, see the
policies and contribution forms [3].

[1] http://www.w3.org/Consortium/Legal/2008/04-testsuite-license
[2] http://www.w3.org/Consortium/Legal/2008/03-bsd-license
[3] http://www.w3.org/2004/10/27-testcases
-->
<html>
<head>
<title>Shadow DOM Test: A_10_04_01</title>
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<link rel="help" href="http://www.w3.org/TR/2013/WD-shadow-dom-20130514/#content-element">
<meta name="assert" content="The content HTML element: fallback content">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script>
test(unit(function (ctx) {

    var d = newRenderedHTMLDocument(ctx);

    d.body.innerHTML =
        '<ul class="cls">' +
            '<li id="li1" class="shadow">1</li>' +
            '<li id="li2" class="shadow2">2</li>' +
            '<li id="li3" class="shadow">3</li>' +
            '<li id="li4">4</li>' +
            '<li id="li5" class="shadow">5</li>' +
            '<li id="li6" class="shadow2">6</li>' +
        '</ul>';

    var host = d.querySelector('.cls');
    //Shadow root to play with
    var s = host.createShadowRoot();

    var div = d.createElement('div');
    div.innerHTML = '' +
        '<ul><content select=".clazz"><span id="spandex">This is fallback content</span></content></ul>';
    s.appendChild(div);

    assert_true(s.querySelector('#spandex').offsetTop > 0, 'Fallback content should be rendered');

}), 'A_10_04_01_T01');



test(unit(function (ctx) {

    var d = newRenderedHTMLDocument(ctx);

    d.body.innerHTML =
        '<ul class="cls">' +
            '<li id="li1" class="shadow">1</li>' +
            '<li id="li2" class="shadow2">2</li>' +
            '<li id="li3" class="shadow">3</li>' +
            '<li id="li4">4</li>' +
            '<li id="li5" class="shadow">5</li>' +
            '<li id="li6" class="shadow2">6</li>' +
        '</ul>';

    var host = d.querySelector('.cls');
    //Shadow root to play with
    var s = host.createShadowRoot();

    var div = d.createElement('div');
    div.innerHTML = '' +
        '<ul><content select=".shadow"><span id="spandex">This is fallback content</span></content></ul>';
    s.appendChild(div);

    assert_equals(s.querySelector('#spandex').offsetTop, 0, 'Fallback content should not be rendered');

}), 'A_10_04_01_T02');
</script>
</body>
</html>
